pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "API para gerenciar persistência de dados usando MongoDB"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi==0.115.6",
    "uvicorn==0.34.0",
    "pydantic==2.10.5",
    "python-dotenv==1.0.1",
    "motor==3.3.2", 
    "pymongo==4.7.2",
    "pip"
]
